% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sumatrix.R
\docType{class}
\name{Submatrix-class}
\alias{Submatrix}
\alias{Submatrix-class}
\alias{show,Submatrix-method}
\alias{subFromEmboss}
\alias{subFromEmboss,Submatrix,character,ANY-method}
\alias{subFromEmboss,Submatrix,character-method}
\title{Class "Submatrix" Substitution Matrix
Amino Acid Substitution Matrices}
\usage{
subFromEmboss(object, path, ...)

\S4method{subFromEmboss}{Submatrix,character}(object, path, ...)

\S4method{show}{Submatrix}(object)

Submatrix()
}
\description{
Class "Submatrix" Substitution Matrix
Amino Acid Substitution Matrices

create Substition matrix from Emboss

create substitution matrix from emboss file

show submatix

construct Submatrix
}

